+++ routes/store.js
@@ -4,6 +4,7 @@ const Store = require("../models/Stores");
 const User = require("../models/Users");
 const Items = require("../models/Items");
 const Comments = require("../models/Comments");
+const auth = require("../authenticate");
 
 const router = express.Router();
 router.use(bodyparser.json());
@@ -14,7 +15,7 @@ router.use(bodyparser.json());
 */
 router
   .route("/")
-  .get(async (req, res, next) => {
+  .get(auth.verifyUser, auth.validateAdmin, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
@@ -25,33 +26,29 @@ router
       next(error);
     }
   })
-  .post(async (req, res, next) => {
+  .post(auth.verifyUser, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
-      let user = await User.findById(req.body.user);
-      if (user === null) {
-        res.json({ status: false, payload: [], error: "User does not exists" });
+      let user = await User.findById(req.user._id);
+      let store = await Store.create(req.body);
+      if (user.store === undefined) {
+        user.store = store._id;
+        let succ = await user.save();
+        res.json({ status: true, payload: store, error: "" });
       } else {
-        let store = await Store.create(req.body);
-        if (user.store === undefined) {
-          user.store = store._id;
-          let succ = await user.save();
-          res.json({ status: true, payload: store, error: "" });
-        } else {
-          res.json({
-            status: false,
-            payload: [],
-            error: "You already have a Store, please use that to sell Items"
-          });
-        }
+        res.json({
+          status: false,
+          payload: [],
+          error: "You already have a Store, please use that to sell Items"
+        });
       }
     } catch (error) {
       res.json({ status: false, payload: [], error: error });
       next(error);
     }
   })
-  .put((req, res, next) => {
+  .put(auth.verifyUser, (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     res.json({
@@ -60,7 +57,7 @@ router
       error: "PUT operation not allowed"
     });
   })
-  .delete(async (req, res, next) => {
+  .delete(auth.verifyUser, auth.validateAdmin, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
@@ -80,7 +77,7 @@ router
 */
 router
   .route("/:storeId")
-  .get(async (req, res, next) => {
+  .get(auth.verifyUser, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
@@ -99,7 +96,7 @@ router
       next(error);
     }
   })
-  .post((req, res, next) => {
+  .post(auth.verifyUser, (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     res.json({
@@ -108,57 +105,77 @@ router
       error: "POST operation not allowed"
     });
   })
-  .put(async (req, res, next) => {
+  .put(auth.verifyUser, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
-      const user = await User.findById(req.body.user);
-      if(user === null) {
-        res.json({status:false, payload:[], error:"User does not exists"});
-      } else {
-        let store = await Store.findByIdAndUpdate(
-          req.params.storeId,
-          { $set: req.body },
-          { new: true }
-        );
-        if(store === null) {
+      if (String(req.user._id) === String(req.body.user)) {
+        const user = await User.findById(req.body.user);
+        if (user === null) {
           res.json({
             status: false,
             payload: [],
-            error: "Store does not exists"
+            error: "User does not exists"
           });
         } else {
-          res.json({ status: true, payload: store, error: "" });
+          let store = await Store.findByIdAndUpdate(
+            req.params.storeId,
+            { $set: req.body },
+            { new: true }
+          );
+          if (store === null) {
+            res.json({
+              status: false,
+              payload: [],
+              error: "Store does not exists"
+            });
+          } else {
+            res.json({ status: true, payload: store, error: "" });
+          }
         }
+      } else {
+        res.json({
+          status: false,
+          payload: [],
+          error: "You cannot modify somebody's store."
+        });
       }
     } catch (error) {
       res.json({ status: false, payload: {}, error: error });
       next(error);
     }
   })
-  .delete(async (req, res, next) => {
+  .delete(auth.verifyUser, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
-      const store = await Store.findByIdAndRemove(req.params.storeId);
-      if (store === null) {
+      if (String(req.user.store) == String(req.params.storeId)) {
+        const store = await Store.findByIdAndRemove(req.params.storeId);
+        if (store === null) {
+          res.json({
+            status: false,
+            payload: [],
+            error: "Store does not exists"
+          });
+        } else {
+          const user = await User.findOne({ store: store._id });
+          user.store = undefined;
+          const succ = await user.save();
+          const items_ = await Items.find({ store: store._id });
+          if (items_.length !== 0) {
+            for (let item of items_) {
+              const comment = await Comments.remove({ item: item._id });
+            }
+            const items = await Items.remove({ store: store._id });
+          }
+          res.json({ status: true, payload: store, error: "" });
+        }
+      } else {
         res.json({
           status: false,
           payload: [],
-          error: "Store does not exists"
+          error: "You cannot delete somebody's store"
         });
-      } else {
-        const user = await User.findOne({ store: store._id });
-        user.store = undefined;
-        const succ = await user.save();
-        const items_ = await Items.find({ store: store._id });
-        if (items_.length !== 0) {
-          for (let item of items_) {
-            const comment = await Comments.remove({ item: item._id });
-          }
-          const items = await Items.remove({ store: store._id });
-        }
-        res.json({ status: true, payload: store, error: "" });
       }
     } catch (error) {
       res.json({ status: false, payload: {}, error: error });
