+++ routes/users.js
@@ -4,6 +4,7 @@ const User = require("../models/Users");
 const Items = require("../models/Items");
 const Store = require("../models/Stores");
 const Comments = require("../models/Comments");
+const auth = require("../authenticate");
 
 const router = express.Router();
 router.use(bodyParser.json());
@@ -14,7 +15,7 @@ router.use(bodyParser.json());
 */
 router
   .route("/")
-  .get(async (req, res, next) => {
+  .get(auth.verifyUser, auth.validateAdmin, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
@@ -25,16 +26,15 @@ router
       next(error);
     }
   })
-  .post(async (req, res, next) => {
+  .post((req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
-    try {
-      let user = await User.create(req.body);
-      res.json({ status: true, payload: user, error: "" });
-    } catch (error) {
-      res.json({ status: false, payload: {}, error: error });
-      next(error);
-    }
+    res.json({
+      status: false,
+      payload: [],
+      error:
+        "You cannot create the user. Please signin with your Google account."
+    });
   })
   .put((req, res, next) => {
     res.setHeader("Content-Type", "application/json");
@@ -45,7 +45,7 @@ router
       error: "PUT operation not allowed"
     });
   })
-  .delete(async (req, res, next) => {
+  .delete(auth.verifyUser, auth.validateAdmin, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
@@ -66,7 +66,7 @@ router
 */
 router
   .route("/:userId")
-  .get(async (req, res, next) => {
+  .get(auth.verifyUser, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
@@ -81,7 +81,7 @@ router
       next(error);
     }
   })
-  .post((req, res, next) => {
+  .post(auth.verifyUser, (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     res.json({
@@ -90,7 +90,7 @@ router
       error: "POST operation not allowed"
     });
   })
-  .put(async (req, res, next) => {
+  .put(auth.verifyUser, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     res.json({
@@ -99,25 +99,37 @@ router
       error: "PUT operation not allowed for Google OAuth"
     });
   })
-  .delete(async (req, res, next) => {
+  .delete(auth.verifyUser, async (req, res, next) => {
     res.setHeader("Content-Type", "application/json");
     res.statusCode = 200;
     try {
-      const user = await User.findByIdAndRemove(req.params.userId);
-      if (user === null) {
-        res.json({ status: false, payload: [], error: "User does not exists" });
-      } else {
-        const store = await Store.findByIdAndRemove(user.store);
-        if (store !== null) {
-          const items_ = await Items.find({ store: store._id });
-          if (items_.length !== 0) {
-            for (let item of items_) {
-              const comments = await Comments.remove({ item: item._id });
+      if (String(req.params.userId) === String(req.user._id)) {
+        const user = await User.findByIdAndRemove(req.params.userId);
+        if (user === null) {
+          res.json({
+            status: false,
+            payload: [],
+            error: "User does not exists"
+          });
+        } else {
+          const store = await Store.findByIdAndRemove(user.store);
+          if (store !== null) {
+            const items_ = await Items.find({ store: store._id });
+            if (items_.length !== 0) {
+              for (let item of items_) {
+                const comments = await Comments.remove({ item: item._id });
+              }
+              const items = await Items.remove({ store: store._id });
             }
-            const items = await Items.remove({ store: store._id });
           }
+          res.json({ status: true, payload: user, error: "" });
         }
-        res.json({ status: true, payload: user, error: "" });
+      } else {
+        res.json({
+          status: false,
+          payload: [],
+          error: "You cannot delete another person's account"
+        });
       }
     } catch (error) {
       res.json({ status: false, payload: {}, error: error });
@@ -129,76 +141,93 @@ router
   @route /api/users/cart/userId
   @desc add and remove an item from the user cart
 */
-router.route("/cart/:userId")
-.post(async (req, res, next) => {
-  res.setHeader("Content-Type","application/json");
-  res.statusCode = 200;
-  try {
-   const user = await User.findById(req.params.userId);
-   const item = await Items.findById(req.body._id);
-   if(user === null) {
-    res.json({
-      status:false,
-      payload:[],
-      error: "User does not exists"
-    });
-   } else if(item === null) {
-    res.json({
-      status:false,
-      payload:[],
-      error: "Item you are trying to add does not exists"
-    });
-   } else {
-    user.cart.push(item);
-    const succ = await user.save();
-    res.json({
-      status:true,
-      payload:item,
-      error:""
-    });
-   }
-  } catch (error) {
-    res.json({
-      status:false,
-      payload:[],
-      error: error
-    });
-  }
-})
-.delete(async (req, res, next) => {
-  res.setHeader("Content-Type","application/json");
-  res.statusCode = 200;
-  try {
-    const user = await User.findById(req.params.userId);
-    const item = await Items.findById(req.params._id);
-    if(user === null) {
+router
+  .route("/:userId/cart/:itemId")
+  .post(auth.verifyUser, async (req, res, next) => {
+    res.setHeader("Content-Type", "application/json");
+    res.statusCode = 200;
+    try {
+      if (String(req.user._id) === String(req.params.userId)) {
+        const user = await User.findById(req.params.userId);
+        const item = await Items.findById(req.params.itemId);
+        if (user === null) {
+          res.json({
+            status: false,
+            payload: [],
+            error: "User does not exists"
+          });
+        } else if (item === null) {
+          res.json({
+            status: false,
+            payload: [],
+            error: "Item you are trying to add does not exists"
+          });
+        } else {
+          user.cart.push(item);
+          const succ = await user.save();
+          res.json({
+            status: true,
+            payload: item,
+            error: ""
+          });
+        }
+      } else {
+        res.json({
+          status: false,
+          payload: [],
+          error: "You cannot add items to other people's cart."
+        });
+      }
+    } catch (error) {
       res.json({
-        status:false,
-        payload:[],
-        error: "User does not exists"
-      });
-    } else if(item === null) {
-      res.json({
-        status:false,
-        payload:[],
-        error: "Item you are trying to delete does not exists in the cart"
-      });
-    } else {
-      user.cart.pull(item._id);
-      const succ = await user.save();
-      res.json({
-        status:true,
-        payload:item,
-        error:""
+        status: false,
+        payload: [],
+        error: error
       });
     }
-  } catch (error) {
-    res.json({
-      status:false,
-      payload:[],
-      error: error
-    });
-  }
-});
+  })
+  .delete(auth.verifyUser, async (req, res, next) => {
+    res.setHeader("Content-Type", "application/json");
+    res.statusCode = 200;
+    try {
+      if (String(req.user._id) === String(req.params.userId)) {
+        const user = await User.findById(req.params.userId);
+        const item = await Items.findById(req.params.itemId);
+        if (user === null) {
+          res.json({
+            status: false,
+            payload: [],
+            error: "User does not exists"
+          });
+        } else if (item === null) {
+          res.json({
+            status: false,
+            payload: [],
+            error: "Item you are trying to delete does not exists in the cart"
+          });
+        } else {
+          user.cart.pull(item._id);
+          const succ = await user.save();
+          res.json({
+            status: true,
+            payload: item,
+            error: ""
+          });
+        }
+      } else {
+        res.json({
+          status: false,
+          payload: [],
+          error: "You cannot delete items from other people's cart"
+        });
+      }
+    } catch (error) {
+      res.json({
+        status: false,
+        payload: [],
+        error: error
+      });
+    }
+  });
 
 module.exports = router;
